app/models/time_off.py
import uuid
from datetime import datetime
from sqlalchemy.dialects.postgresql import UUID
from ..extensions import db

class TherapistTimeOff(db.Model):
    __tablename__ = "therapist_time_off"

    id = db.Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    start_at = db.Column(db.DateTime(timezone=True), nullable=False)
    end_at = db.Column(db.DateTime(timezone=True), nullable=False)
    reason = db.Column(db.Text, nullable=True)
    created_at = db.Column(db.DateTime(timezone=True), nullable=False, default=datetime.utcnow)
