requirements.txt
Flask==3.0.3
python-dotenv==1.0.1


pytest==8.3.2


gunicorn==22.0.0


Flask-SQLAlchemy
psycopg2-binary
python-dotenv